requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect==1.2.0
discord.py==2.2.2
fake-useragent==1.1.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
lxml==4.9.2
multidict==6.0.4
parse==1.19.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dotenv==1.0.0
requests==2.28.2
requests-html==0.10.0
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
yarl==1.8.2
zipp==3.15.0
